refactor: Move ngram functions into separate ngram module
